# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jef Bot API",
    "version": "1.0.0",
    "summary": "AI chat moderation for Discord servers, game chat and any app.",
    "description": "The /v1 API of Jef Bot. Authenticate with `Authorization: Bearer <key>`: account keys (jef_acct_…) can do everything on the account; app keys (jef_app_…) can only moderate and read one app (plus warn/pardon its users). Errors are `{ \"error\": { \"code\", \"message\" } }`. Timestamps are ISO 8601 UTC. Guides: /docs/agents.html, /docs/quickstart-game.html. Full text for LLMs: /llms-full.txt.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This Jef Bot server"
    }
  ],
...
  "security": [
    {
      "accountKey": []
    },
    {
      "appKey": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "tags": [
          "Public"
        ],
        "summary": "Public plans and their limits",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "Plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plans": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Plan"
                      }
                    }
                  },
                  "required": [
                    "plans"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          }
        }
      }
    },
    "/v1/account": {
      "get": {
        "operationId": "getAccount",
        "tags": [
          "Account"
        ],
        "summary": "The calling account, its limits and usage this month",
        "security": [
          {
            "accountKey": []
          },
          {
            "appKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          }
        }
      }
    },
    "/v1/keys": {
      "get": {
        "operationId": "listAccountKeys",
        "tags": [
          "Account keys"
        ],
        "summary": "List account keys",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Key"
                      }
                    }
                  },
                  "required": [
                    "keys"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          }
        }
      },
      "post": {
        "operationId": "createAccountKey",
        "tags": [
          "Account keys"
        ],
        "summary": "Create an account key",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 60
                  }
                },
                "required": []
              },
              "example": {
                "name": "CI setup agent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created; the secret is shown only once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewKey"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          }
        }
      }
    },
    "/v1/keys/{keyId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/keyId"
        }
      ],
      "delete": {
        "operationId": "deleteAccountKey",
        "tags": [
          "Account keys"
        ],
        "summary": "Delete an account key",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      }
    },
    "/v1/apps": {
      "get": {
        "operationId": "listApps",
        "tags": [
          "Apps"
        ],
        "summary": "List apps",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Apps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apps": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/App"
                      }
                    }
                  },
                  "required": [
                    "apps"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          }
        }
      },
      "post": {
        "operationId": "createApp",
        "tags": [
          "Apps"
        ],
        "summary": "Create an app (with rules, webhook and first app key in one call)",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppRequest"
              },
              "example": {
                "name": "My Game",
                "kind": "api",
                "rules": [
                  {
                    "text": "No begging other players for items or currency",
                    "severity": "low"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAppResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          }
        }
      }
    },
    "/v1/apps/{appId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/appId"
        }
      ],
      "get": {
        "operationId": "getApp",
        "tags": [
          "Apps"
        ],
        "summary": "Get an app",
        "security": [
          {
            "accountKey": []
          },
          {
            "appKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "App",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "app": {
                      "$ref": "#/components/schemas/App"
                    }
                  },
                  "required": [
                    "app"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      },
      "patch": {
        "operationId": "updateApp",
        "tags": [
          "Apps"
        ],
        "summary": "Update name, webhook URL or Discord settings",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppRequest"
              },
              "example": {
                "webhookUrl": "https://game.example.com/hooks/jef"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "app": {
                      "$ref": "#/components/schemas/App"
                    }
                  },
                  "required": [
                    "app"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      },
      "delete": {
        "operationId": "deleteApp",
        "tags": [
          "Apps"
        ],
        "summary": "Delete an app and all its data",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      }
    },
    "/v1/apps/{appId}/webhook-secret": {
      "parameters": [
        {
          "$ref": "#/components/parameters/appId"
        }
      ],
      "post": {
        "operationId": "rotateWebhookSecret",
        "tags": [
          "Apps"
        ],
        "summary": "Rotate the webhook signing secret",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "New secret",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "webhookSecret": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "webhookSecret"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      }
    },
...
    "/v1/apps/{appId}/keys/{keyId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/appId"
        },
        {
          "$ref": "#/components/parameters/keyId"
        }
      ],
      "delete": {
        "operationId": "deleteAppKey",
        "tags": [
          "App keys"
        ],
        "summary": "Delete an app key",
        "security": [
          {
            "accountKey": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "accountKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jef_acct_…",
        "description": "Account key: full access to the account."
      },
      "appKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jef_app_…",
        "description": "App key: moderation and reads for one app."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jef_session",
        "description": "Dashboard session after Google sign-in. Changes must be same-origin JSON."
      }
    },
    "schemas": {
...
      "Plan": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PlanLimits"
          },
          {
            "type": "object",
            "properties": {
              "purchasable": {
                "type": "boolean",
                "description": "Whether this server can sell the plan (Free is always true)."
              }
            },
            "required": [
              "purchasable"
            ]
          }
        ]
      },
...
    },
    "responses": {
...
      "E401": {
        "description": "unauthorized (no credentials) or invalid_api_key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NoContent": {
        "description": "Done (no body)."
      }
    },
...
  }
}